BigIntegerField

+++ backend/server/apps/endpoints/migrations/0001_initial.py
@@ -26,7 +26,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('measurement_id', models.IntegerField()),
-                ('timestamp', models.IntegerField()),
+                ('timestamp', models.BigIntegerField()),
                 ('measurement_type', models.CharField(max_length=3)),
                 ('limp_type', models.CharField(max_length=1)),
                 ('limp_side', models.CharField(max_length=1)),
